fix council
