+++ products/council/analysis.py
@@ -1,0 +1,28 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def analyze_responses(responses, confidences, expertise_scores):
+    # This is a placeholder function. You can implement more sophisticated
+    # analysis here based on the responses, confidences, and expertise scores.
+    try:
+        analysis = {
+            "total_responses": len(responses),
+            "average_confidence": (
+                sum(confidences.values()) / len(confidences)
+                if confidences
+                else 0
+            ),
+            "highest_expertise": (
+                max(expertise_scores.values()) if expertise_scores else 0
+            ),
+        }
+        return analysis
+    except Exception as e:
+        logger.error(f"Error in analyze_responses: {e}")
+        return {
+            "total_responses": 0,
+            "average_confidence": 0,
+            "highest_expertise": 0,
+        }
